requirements.txt
# Core ML
torch
sentence-transformers
faiss-cpu
transformers
peft

# LangChain / LangGraph
langchain
langchain-community
langgraph

# Data & Storage
pandas
requests
sqlite-utils

# Validation
pydantic

# UI
streamlit

# Evaluation
scikit-learn
numpy

# Dev
pytest
python-dotenv
jupyter
